requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

Django==1.11.5
django-extensions
pinax-theme-bootstrap==2.0.4
pinax-theme-bootstrap-account==1.0b2
django-appconf==1.0.2
webtest
django-webtest
pinax-notifications
toposort
django-cors-headers
easy_thumbnails
rdflib
rdflib-jsonld
djangorestframework
jsonrpclib
django-contrib-comments
html5lib>0.99999999
bleach==2.0.0
django-simple-captcha
django-mptt
pytz
graphql-core==1.1
graphene-django==1.3
graphene==1.4.1
PyJWT
six
selenium
mock
requests

